+++ src/Stack/Options/TestParser.hs
@@ -19,13 +19,13 @@ testOptsParser :: Bool -> Parser TestOptsMonoid
 testOptsParser hide0 = TestOptsMonoid
   <$> firstBoolFlagsTrue
         "rerun-tests"
-        "running already successful tests."
+        "running already successful test suites."
         hide
   <*> fmap concat (many (argsOption
         (  long "test-arguments"
         <> long "ta"
         <> metavar "TEST_ARGS"
-        <> help "Arguments passed in to the test suite program."
+        <> help "Arguments passed to the test suites."
         <> hide
         )))
   <*> optionalFirstFalse (flag' True
@@ -33,11 +33,10 @@ testOptsParser hide0 = TestOptsMonoid
         <> help "Generate a code coverage report."
         <> hide
         ))
-  <*> optionalFirstFalse (flag' True
-        (  long "no-run-tests"
-        <> help "Disable running of tests. (Tests will still be built.)"
-        <> hide
-        ))
+  <*> firstBoolFlagsTrue
+        "run-tests"
+        "running of targeted test suites."
+        hide
   <*> optionalFirst (option (fmap Just auto)
         (  long "test-suite-timeout"
         <> help "Maximum test suite run time in seconds."
@@ -45,7 +44,7 @@ testOptsParser hide0 = TestOptsMonoid
         ))
   <*> firstBoolFlagsTrue
         "tests-allow-stdin"
-        "allow standard input in test executables."
+        "allow standard input in test suites."
         hide
  where
   hide = hideMods hide0